Se ajusta endpoint de registrar usuario y se agrega endpoint de eliminar usuario

+++ src/main/java/com/alkemy/wallet/dto/RoleDTO.java
@@ -1,6 +1,6 @@
 package com.alkemy.wallet.dto;
 
-import com.alkemy.wallet.model.entity.RoleEnum;
+import com.alkemy.wallet.model.RoleEnum;
 import lombok.*;
 
 @ToString
@@ -9,7 +9,7 @@ import lombok.*;
 @Getter
 @Setter
 public class RoleDTO {
-    private int id;
+    private Long id;
     private RoleEnum name;
     private String description;
 }
